GH-3913: Refactoring
