+++ projects/storefrontlib/src/components/asm/asm-root/asm-root.component.html
@@ -1,1 +1,3 @@
-<cx-asm-main-ui *cxFeature="'asm'"></cx-asm-main-ui>
+<ng-container *cxFeature="'asm'"
+  ><cx-asm-main-ui *ngIf="(asmUi$ | async)?.visible"></cx-asm-main-ui
+></ng-container>
